* libmutil: BSD patch for libmutil from Erik Norgaard. Thanks.
git-svn-id: svn://svn.minisip.org/minisip/trunk@866 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/source/Thread.cxx
@@ -176,7 +176,7 @@ int Thread::createThread(void f()){
 	pthread_t threadHandle;
 //	cerr << "Running createThread"<< endl;
 	pthread_create(&threadHandle, NULL, LinuxStaticThreadStarter, (void*)f);
-	return threadHandle;
+	return (int)threadHandle;
 #endif
 
 }
@@ -211,7 +211,7 @@ int Thread::createThread(void *f(void*), void *arg){
 	pthread_t threadHandle;
 //	cerr << "Running createThread"<< endl;
 	pthread_create(&threadHandle, NULL, f, arg);
-	return threadHandle;
+	return (int)threadHandle;
 #endif
 
 }
